requirements.txt
pytorch
transformers
numpy
scikit-learn